Fix wasteful packing of unsigned integers required by Java but not C#.
Store TIFF tags (including Exif) using C#'s richer primitive types.

Java only has signed integral primitives, yet TIFF specifies both signed and unsigned values.
In Java, metadata-extractor had to double the size of values to store unsigned values correctly.
In C# we have both signed and unsigned versions of integral primitives, so we use them here.

+++ Com.Drew.Tests/Com/drew/metadata/exif/NikonType1MakernoteTest.cs
@@ -87,8 +87,8 @@ namespace Com.Drew.Metadata.Exif
             Assert.AreEqual(string.Empty, _nikonDirectory.GetString(NikonType1MakernoteDirectory.TagUnknown2));
             Assert.AreEqual(0, _nikonDirectory.GetDouble(NikonType1MakernoteDirectory.TagDigitalZoom), 0.0001);
             Assert.AreEqual(0, _nikonDirectory.GetInt(NikonType1MakernoteDirectory.TagConverter));
-            long[] unknown3 = (long[])_nikonDirectory.GetObject(NikonType1MakernoteDirectory.TagUnknown3);
-            long[] expected = new long[] { 0, 0, 16777216, 0, 2685774096L, 0, 34833, 6931, 16178, 4372, 4372, 3322676767L, 3373084416L, 15112, 0, 0, 1151495, 252903424, 17, 0, 0, 844038208, 55184128, 218129428, 1476410198, 370540566, 4044363286L, 16711749
+            uint[] unknown3 = (uint[])_nikonDirectory.GetObject(NikonType1MakernoteDirectory.TagUnknown3);
+            uint[] expected = new uint[] { 0, 0, 16777216, 0, 2685774096, 0, 34833, 6931, 16178, 4372, 4372, 3322676767, 3373084416, 15112, 0, 0, 1151495, 252903424, 17, 0, 0, 844038208, 55184128, 218129428, 1476410198, 370540566, 4044363286, 16711749
                 , 204629079, 1729 };
             Assert.IsNotNull(unknown3);
             Assert.AreEqual(expected.Length, unknown3.Length);
